Added Space Invaders Game into main.py file.
Changed paths of recourses in the Space Invaders game to work in the main.py file. Made the whole 'Space_Invaders.py' into a function so it can be called from main.py. Added a call for said function in Top Left button on the main screen. Added Space Invaders thumbnail to the main screen in the Top Left button.

+ Updated .gitignore.